+++ jobbergate-cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 This file keeps track of all notable changes to jobbergate-cli
 
 ## Unreleased
+- Added command to render an application into a job script locally [ASP-3326]
 
 ## 4.3.0 -- 2024-02-14